// mantalents/src/interfaces/postMovimiento.tsx
// Generated by https://quicktype.io

export interface PostMovimiento {
    codigo:    string;
    fecha:     string;
    cliente:   string;
    subtotal:  number;
    iva:       number;
    descuento: number;
    total:     number;
    productos: PostProducto[];
}

export interface PostProducto {
    item:     string;
    cantidad: number;
    precio:   number;
    total?: number;
}